refactor(backend): remove .v2 suffix from migrated handlers
Big bang replacement - .v2 naming no longer needed:
- Deleted old handlers (register.ts, login.ts, refresh.ts)
- Renamed .v2 files to canonical names
- Moved tests from __tests__/ to auth/ directory
- Updated import paths in all test files

All handlers now use Middy middleware exclusively.
Tests still passing: 14/27 (13 require LocalStack).

Files renamed:
- register.v2.ts → register.ts
- login.v2.ts → login.ts
- refresh.v2.ts → refresh.ts
- *.v2.test.ts → *.test.ts

Git provides version control - no need for file-level versioning.
