+++ packages/backend/src/handlers/auth/register.ts
@@ -1,31 +1,46 @@
-import { RegisterRequestSchema } from '@social-media-app/shared';
-import { compose } from '../../infrastructure/middleware/compose.js';
-import { withErrorHandling } from '../../infrastructure/middleware/withErrorHandling.js';
-import { withLogging } from '../../infrastructure/middleware/withLogging.js';
-import { withValidation } from '../../infrastructure/middleware/withValidation.js';
-import { withServices } from '../../infrastructure/middleware/withServices.js';
-import { successResponse } from '../../utils/responses.js';
-
 /**
- * Lambda handler for user registration
+ * Register Handler (Middy Version)
  *
- * Creates a new user account with email, username, and password.
+ * Migrated from custom middleware composition to Middy.
+ * Creates new user accounts with email, username, and password.
  *
  * @route POST /auth/register
- * @middleware withErrorHandling - Converts errors to HTTP responses (including conflict errors)
- * @middleware withLogging - Structured logging with correlation IDs
- * @middleware withValidation - Validates request body against RegisterRequestSchema
- * @middleware withServices - Injects authService into context
  */
-export const handler = compose(
-  withErrorHandling(),
-  withLogging(),
-  withValidation(RegisterRequestSchema),
-  withServices(['authService']),
-  async (_event, context) => {
-    // Business logic only - middleware handles validation, logging, and errors
-    // Non-null assertion safe: withServices middleware guarantees authService exists
-    const response = await context.services!.authService.register(context.validatedInput);
-    return successResponse(201, response);
+
+import { RegisterRequestSchema, type RegisterRequest } from '@social-media-app/shared'
+import { authService } from '../../utils/services.js'
+import { createHandler } from '../../infrastructure/middleware-v2/index.js'
+import type { APIGatewayProxyHandlerV2 } from 'aws-lambda'
+
+/**
+ * Handler implementation - pure business logic
+ *
+ * No middleware concerns - Middy handles:
+ * - JSON parsing
+ * - Validation
+ * - Error handling
+ * - Header normalization
+ */
+const registerHandler: APIGatewayProxyHandlerV2 = async (event) => {
+  // Type-safe access to validated input
+  const request = event.validatedBody as RegisterRequest
+
+  // Business logic only - delegate to authService
+  const response = await authService.register(request)
+
+  return {
+    statusCode: 201,
+    headers: {
+      'Content-Type': 'application/json',
+      'Access-Control-Allow-Origin': '*'
+    },
+    body: JSON.stringify(response)
   }
-);
+}
+
+/**
+ * Export Middy-wrapped handler with middleware stack
+ */
+export const handler = createHandler(registerHandler, {
+  validation: RegisterRequestSchema
+})
